backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gregbi <grigorybil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
python-dotenv = "^0.19.2"
PyJWT = "^2.3.0"
ormar = "^0.11.0"
aioyagmail = {extras = ["all"], version = "^0.0.4"}
asyncpg = "^0.25.0"
psycopg2 = "^2.9.3"
websockets = "^10.3"
python-socketio = "^5.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
